Refactor packages and add documentations
