+++ balloon/src/main/java/com/skydoves/balloon/OnBalloonClickListener.kt
@@ -20,5 +20,7 @@ import android.view.View
 
 /** Interface definition for a callback to be invoked when a balloon view is clicked. */
 interface OnBalloonClickListener {
+
+  /** invoked when the [Balloon] is clicked. */
   fun onBalloonClick(view: View)
 }
