+++ balloon/src/main/java/com/skydoves/balloon/OnBalloonDismissListener.kt
@@ -18,5 +18,7 @@ package com.skydoves.balloon
 
 /** Interface definition for a callback to be invoked when a balloon view is dismissed. */
 interface OnBalloonDismissListener {
+
+  /** invoked when the [Balloon] is dismissed. */
   fun onBalloonDismiss()
 }
