+++ balloon/src/main/java/com/skydoves/balloon/OnBalloonOutsideTouchListener.kt
@@ -21,5 +21,7 @@ import android.view.View
 
 /** Interface definition for a callback to be invoked when touched on outside of the balloon popup. */
 interface OnBalloonOutsideTouchListener {
+
+  /** invoked when the outside of the [Balloon] is touched. */
   fun onBalloonOutsideTouch(view: View, event: MotionEvent)
 }
