+++ balloon/src/main/res/values/attrs.xml
@@ -1,14 +1,14 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
-    <declare-styleable name="VectorTextView">
-        <attr name="drawableLeft" format="reference" />
-        <attr name="drawableRight" format="reference" />
-        <attr name="drawableTop" format="reference" />
-        <attr name="drawableBottom" format="reference" />
-        <attr name="drawablePadding" format="dimension" />
-        <attr name="drawableTintColor" format="color" />
-        <attr name="drawableWidth" format="dimension" />
-        <attr name="drawableHeight" format="dimension" />
-        <attr name="drawableSquareSize" format="dimension" />
-    </declare-styleable>
+  <declare-styleable name="VectorTextView">
+    <attr name="drawableLeft" format="reference" />
+    <attr name="drawableRight" format="reference" />
+    <attr name="drawableTop" format="reference" />
+    <attr name="drawableBottom" format="reference" />
+    <attr name="drawablePadding" format="dimension" />
+    <attr name="drawableTintColor" format="color" />
+    <attr name="drawableWidth" format="dimension" />
+    <attr name="drawableHeight" format="dimension" />
+    <attr name="drawableSquareSize" format="dimension" />
+  </declare-styleable>
 </resources>
